Move the fix to `expo-modules-core`

+++ packages/expo-video/ios/VideoModule.swift
@@ -142,14 +142,11 @@ public final class VideoModule: Module {
         return player.pointer.currentItem?.duration.seconds ?? 0
       }
 
-      // JS sends a Double when calling the set function, which sometimes causes the `as?` conversion
-      // in expo-modules-core to fail when playbackRate is defined as a Float.
-      // AVPlayer uses Float for rate, so cast to Double only for communication with JS
-      Property("playbackRate") { player -> Double in
-        return Double(player.playbackRate)
+      Property("playbackRate") { player -> Float in
+        return player.playbackRate
       }
-      .set { (player, playbackRate: Double) in
-        player.playbackRate = Float(playbackRate)
+      .set { (player, playbackRate: Float) in
+        player.playbackRate = playbackRate
       }
 
       Property("isLive") { player -> Bool in
@@ -174,12 +171,11 @@ public final class VideoModule: Module {
         return player.status
       }
 
-      // Reason for type casting same as playbackRate
-      Property("volume") { player -> Double in
-        return Double(player.volume)
+      Property("volume") { player -> Float in
+        return player.volume
       }
-      .set { (player, volume: Double) in
-        player.volume = Float(volume)
+      .set { (player, volume: Float) in
+        player.volume = volume
       }
 
       Function("play") { player in
